Move LCTContainer to separate files
