+++ L1Trigger/CSCTriggerPrimitives/src/CSCUpgradeMotherboard.cc
@@ -1,49 +1,5 @@
 #include "L1Trigger/CSCTriggerPrimitives/interface/CSCUpgradeMotherboard.h"
 
-CSCUpgradeMotherboard::LCTContainer::LCTContainer(unsigned int trig_window_size)
-    : match_trig_window_size_(trig_window_size) {}
-
-CSCCorrelatedLCTDigi& CSCUpgradeMotherboard::LCTContainer::operator()(int bx, int match_bx, int lct) {
-  return data[bx][match_bx][lct];
-}
-
-void CSCUpgradeMotherboard::LCTContainer::getTimeMatched(const int bx, std::vector<CSCCorrelatedLCTDigi>& lcts) const {
-  for (unsigned int mbx = 0; mbx < match_trig_window_size_; mbx++) {
-    for (int i = 0; i < CSCConstants::MAX_LCTS_PER_CSC; i++) {
-      // consider only valid LCTs
-      if (not data[bx][mbx][i].isValid())
-        continue;
-
-      // remove duplicated LCTs
-      if (std::find(lcts.begin(), lcts.end(), data[bx][mbx][i]) != lcts.end())
-        continue;
-
-      lcts.push_back(data[bx][mbx][i]);
-    }
-  }
-}
-
-void CSCUpgradeMotherboard::LCTContainer::getMatched(std::vector<CSCCorrelatedLCTDigi>& lcts) const {
-  for (int bx = 0; bx < CSCConstants::MAX_LCT_TBINS; bx++) {
-    std::vector<CSCCorrelatedLCTDigi> temp_lcts;
-    CSCUpgradeMotherboard::LCTContainer::getTimeMatched(bx, temp_lcts);
-    lcts.insert(std::end(lcts), std::begin(temp_lcts), std::end(temp_lcts));
-  }
-}
-
-void CSCUpgradeMotherboard::LCTContainer::clear() {
-  // Loop over all time windows
-  for (int bx = 0; bx < CSCConstants::MAX_LCT_TBINS; bx++) {
-    // Loop over all matched trigger windows
-    for (unsigned int mbx = 0; mbx < match_trig_window_size_; mbx++) {
-      // Loop over all stubs
-      for (int i = 0; i < CSCConstants::MAX_LCTS_PER_CSC; i++) {
-        data[bx][mbx][i].clear();
-      }
-    }
-  }
-}
-
 CSCUpgradeMotherboard::CSCUpgradeMotherboard(unsigned endcap,
                                              unsigned station,
                                              unsigned sector,
